Issue #127 improve player card selection - modified player card selection screen to be a modal instead of redirecting to a different page - removed cardselection.jsx - removed cardselection from routes
