+++ src/routes.js
@@ -3,7 +3,6 @@ import Game from './components/GamePages/game';
 import Login from './components/LoginPages/login';
 import PlayerWaitinglists from './components/PlayerPages/playerWaitinglists';
 import GameWaitinglists from './components/GamePages/gameWaitinglists';
-import CardSelection from './components/GamePages/cardSelection';
 import GameHistory from './components/HistoryPages/gameHistory';
 import GameHistories from './components/HistoryPages/gameHistories'
 import PlayerGameHistories from './components/HistoryPages/playerGameHistories';
@@ -44,11 +43,6 @@ export const routes = [
     exact: true,
     path: '/games/:id/waitinglists',
     component: GameWaitinglists
-  },
-  {
-    exact: true,
-    path: '/games/:game_id/player_sessions/:id/edit',
-    component: CardSelection
   },
   {
     exact: true,
